requirements.txt
contourpy==1.3.0
cycler==0.12.1
dill==0.3.9
filelock==3.13.1
fonttools==4.56.0
fsspec==2024.6.1
h5py==3.13.0
importlib_resources==6.5.2
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.4
mpmath==1.3.0
multiprocess==0.70.17
networkx==3.2.1
numpy==1.26.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-ignite==0.5.2
pytz==2025.2
SALib==1.5.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
typing_extensions==4.12.2
tzdata==2025.2
zipp==3.21.0
